+++ src/test/java/com/sparta/hhplusconcert/point/usecase/PaySeatServiceTest.java
@@ -1,22 +1,22 @@
-package com.sparta.hhplusconcert.usecase.point;
+package com.sparta.hhplusconcert.point.usecase;
 
 import static org.mockito.Mockito.*;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import com.sparta.hhplusconcert.domain.concert.ReservationStatus;
-import com.sparta.hhplusconcert.domain.concert.SeatStatus;
-import com.sparta.hhplusconcert.domain.concert.entity.ConcertReservationEntity;
-import com.sparta.hhplusconcert.domain.concert.entity.ConcertSeatEntity;
-import com.sparta.hhplusconcert.domain.point.PointTransactionType;
-import com.sparta.hhplusconcert.domain.point.entity.PaymentEntity;
-import com.sparta.hhplusconcert.domain.point.entity.PointHistoryEntity;
-import com.sparta.hhplusconcert.domain.point.entity.UserEntity;
-import com.sparta.hhplusconcert.infra.concert.ConcertReservationRepositoryImpl;
-import com.sparta.hhplusconcert.infra.concert.ConcertSeatRepositoryImpl;
-import com.sparta.hhplusconcert.infra.point.PaymentRepositoryImpl;
-import com.sparta.hhplusconcert.infra.point.PointHistoryRepositoryImpl;
-import com.sparta.hhplusconcert.infra.point.UserRepositoryImpl;
+import com.sparta.hhplusconcert.concert.domain.ReservationStatus;
+import com.sparta.hhplusconcert.concert.domain.SeatStatus;
+import com.sparta.hhplusconcert.concert.domain.entity.ConcertReservationEntity;
+import com.sparta.hhplusconcert.concert.domain.entity.ConcertSeatEntity;
+import com.sparta.hhplusconcert.point.domain.PointTransactionType;
+import com.sparta.hhplusconcert.point.domain.entity.PaymentEntity;
+import com.sparta.hhplusconcert.point.domain.entity.PointHistoryEntity;
+import com.sparta.hhplusconcert.point.domain.entity.UserEntity;
+import com.sparta.hhplusconcert.concert.infra.ConcertReservationRepositoryImpl;
+import com.sparta.hhplusconcert.concert.infra.ConcertSeatRepositoryImpl;
+import com.sparta.hhplusconcert.point.infra.PaymentRepositoryImpl;
+import com.sparta.hhplusconcert.point.infra.PointHistoryRepositoryImpl;
+import com.sparta.hhplusconcert.point.infra.UserRepositoryImpl;
 import java.util.UUID;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
